refactor: move user entity and repository to common package
