+++ app/src/main/kotlin/com/example/common/security/DomainUserDetailsService.kt
@@ -1,6 +1,6 @@
 package com.example.common.security
 
-import com.example.users.UserRepository
+import com.example.common.repositories.UserRepository
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.core.userdetails.User
 import org.springframework.security.core.userdetails.UserDetails
@@ -18,7 +18,7 @@ class DomainUserDetailsService(val userRepository: UserRepository) : UserDetails
             ?.let { createSpringSecurityUser(it) }
             ?: throw UsernameNotFoundException("user account $username NOT found OR disabled")
 
-    private fun createSpringSecurityUser(domainUser: com.example.users.DomainUser): User =
+    private fun createSpringSecurityUser(domainUser: com.example.common.entities.DomainUser): User =
         User(
             domainUser.username,
             domainUser.password,
