+++ app/src/main/kotlin/com/example/users/AccountService.kt
@@ -1,9 +1,12 @@
 package com.example.users
 
+import com.example.common.entities.DomainUser
+import com.example.common.entities.UserStatus
 import com.example.common.events.AccountActivatedEvent
 import com.example.common.events.AccountCreatedEvent
 import com.example.common.events.PasswordChangedEvent
 import com.example.common.events.PasswordResetRequested
+import com.example.common.repositories.UserRepository
 import com.example.common.security.SecurityUtils
 import com.example.common.web.ApplicationException
 import jakarta.validation.constraints.Email
