+++ app/src/main/kotlin/com/example/users/AccountsResource.kt
@@ -1,5 +1,6 @@
 package com.example.users
 
+import com.example.common.entities.DomainUser
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.Size
 import org.openapitools.api.AccountsApi
