+++ app/src/test/kotlin/com/example/users/AccountIT.kt
@@ -2,6 +2,7 @@ package com.example.users
 
 import com.example.common.JPATestConfiguration
 import com.example.common.events.PasswordChangedEvent
+import com.example.common.repositories.UserRepository
 import com.example.users.SignInIT.Companion.DUMMY_EMAIL
 import com.example.users.SignInIT.Companion.DUMMY_PASSWORD
 import com.example.users.SignInIT.Companion.DUMMY_USERNAME
