+++ app/src/test/kotlin/com/example/users/SignInIT.kt
@@ -1,6 +1,8 @@
 package com.example.users
 
 import com.example.common.JPATestConfiguration
+import com.example.common.entities.DomainUser
+import com.example.common.repositories.UserRepository
 import com.fasterxml.jackson.databind.ObjectMapper
 import java.util.stream.Stream
 import org.junit.jupiter.api.BeforeEach
