+++ app/src/test/kotlin/com/example/users/SignUpIT.kt
@@ -1,7 +1,10 @@
 package com.example.users
 
 import com.example.common.JPATestConfiguration
+import com.example.common.entities.DomainUser
+import com.example.common.entities.UserStatus
 import com.example.common.events.AccountCreatedEvent
+import com.example.common.repositories.UserRepository
 import com.fasterxml.jackson.databind.ObjectMapper
 import java.util.stream.Stream
 import kotlin.test.junit5.JUnit5Asserter
