+++ app/src/test/kotlin/com/example/users/UsersIT.kt
@@ -1,6 +1,8 @@
 package com.example.users
 
 import com.example.common.JPATestConfiguration
+import com.example.common.entities.DomainUser
+import com.example.common.repositories.UserRepository
 import com.example.users.SignInIT.Companion.DUMMY_EMAIL
 import com.example.users.SignInIT.Companion.DUMMY_PASSWORD
 import com.example.users.SignInIT.Companion.DUMMY_USERNAME
