+++ app/src/test/kotlin/com/example/users/UsersRepositoryTest.kt
@@ -3,6 +3,8 @@ package com.example.users
 import com.example.common.DEFAULT_TEST_AUDITOR
 import com.example.common.JPATestConfiguration
 import com.example.common.configs.AuthorityConstants
+import com.example.common.entities.DomainUser
+import com.example.common.repositories.UserRepository
 import java.time.Instant
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
